refactor: reorder some classes packages after merge with develop
