+++ identity/src/main/kotlin/net/thechance/identity/api/mapper/UserMapper.kt
@@ -1,7 +1,7 @@
 package net.thechance.identity.api.mapper
 
-import net.thechance.identity.api.dto.ManagedUserResponse
-import net.thechance.identity.api.dto.PageResponse
+import net.thechance.identity.api.dto.profile.ManagedUserResponse
+import net.thechance.identity.api.dto.profile.PageResponse
 import net.thechance.identity.api.dto.profile.ProfileResponse
 import net.thechance.identity.entity.User
 import org.springframework.data.domain.Page
